requirements.txt
fastapi 
google-generativeai 
pydantic
uvicorn
mysql-connector-python
python-dotenv
sqlalchemy

